+++ igdectk/auth/ldap.py
@@ -23,6 +23,53 @@ class LdapAuthenticationBackend(ModelBackend):
 
     """
         Authenticates against settings.AUTH_USER_MODEL and LDAP.
+
+        Here is the list of LDAP settings and options
+        ---------------------------------------------
+
+        The Django settings must contain a dict named LDAPS like :
+
+            LDAPS = {
+                'default': {
+                    'HOST': "<ldap.hostname>",
+                    'USER_DN': "<uid=%s<,ou=...,dc=...,...>",
+                    'OPTIONS': {
+                        'auto_add_user': <Boolean>,
+                        'search_filter': "(<field=value,...)",
+                        'email_fields': [<'field1'>, <...>] or None,
+                        'state_fields': {<'field1'>: <'active_state>', <...>: <...>, ...} or None,
+                        'firstname_fields': [<'field1'>, <...>] or None,
+                        'lastname_fields': [<'field1'>, <...>] or None,
+                    }
+                },
+            }
+
+            - 'default' : dict, used by this backend. Any others entry are ignored.
+            - 'HOST' : str, contains the ldap hostname (with port if necessary).
+            - 'USER_DN' : str, cotains the user domain name as a string. The parameter representing
+                the user identifier must be equal to '%s' because this field is completed by the username
+                coming from the authentication method.
+            - 'OPTIONS' : dict, containing a set of options.
+                - 'auto_add_user' : Boolean, True means users that does not exists into
+                    the user table are automaticaly created. The next parameters are relavant
+                    if this setting is True.
+                - 'search_filter' : str, LDAP search filter is a list wrapped by parenthesis,
+                    containing pairs of fields name and value separate by a equal sign (=), and each
+                    pair are separate by a comma (,).
+                - 'email_fields' : list of str, Defines a list of candidats, in order of priority,
+                    containing an email adresse. Or not defined if email field is not wanted.
+                - 'state_fields' ; dict of str pairs, Defines a dict of candidats, in order of priority,
+                    containing the user status (active, or not) and its value. In others words,
+                    the key contains a name of attribute, and its value is the value meaning
+                    the account is in its active state.
+                    Or None if status field should be not checked.
+                - 'firstname_fields' : list of str, Defines a list of candidats, in order of priority,
+                    containing a first name. Or not defined if first name field is not wanted.
+                - 'lastname_fields' : list of str, Defines a list of candidats, in order of priority,
+                    containing a last name. Or not defined if last name field is not wanted.
+
+        If state fields return a disabled account, then the authentication returns None.
+        If email fields returns empty email then the newly created account will have an empty email.
     """
 
     def authenticate(self, username=None, password=None, **kwargs):
@@ -43,30 +90,37 @@ class LdapAuthenticationBackend(ModelBackend):
             # auto creation of non registred users
             if options.get('auto_add_user', True):
                 email = ''
+                firstname = ''
+                lastname = ''
 
-                get_email = options.get('get_email', False)
-                get_state = options.get('get_state', False)
+                state_fields = options.get('state_fields', {})
+                email_fields = options.get('email_fields', [])
+                firstname_fields = options.get('firstname_fields', [])
+                lastname_fields = options.get('lastname_fields', [])
+                search_filter = options.get('search_filter', '()')
 
-                if get_email or get_state:
-                    email_fields = options.get('email_fields', [])
-                    state_fields = options.get('state_fields', {})
-                    search_filter = options.get('search_filter', '()')
+                # process an LDAP query for the uid
+                srv = ldap3.Server(settings.LDAPS['default']['HOST'])
+                conn = ldap3.Connection(srv, auto_bind=True, client_strategy=ldap3.SYNC)
 
-                    # process an LDAP query for the uid
-                    srv = ldap3.Server(settings.LDAPS['default']['HOST'])
-                    conn = ldap3.Connection(srv, auto_bind=True, client_strategy=ldap3.SYNC)
+                attributes = ['uid']
 
-                    attributes = ['uid']
+                if email_fields:
+                    attributes.extend(email_fields)
 
-                    if get_email:
-                        attributes.extend(email_fields)
+                if firstname_fields:
+                    attributes.extend(firstname_fields)
 
-                    if get_state:
-                        attributes.extend(state_fields.keys())
-                        active_user = False
-                    else:
-                        active_user = True
+                if lastname_fields:
+                    attributes.extend(lastname_fields)
 
+                if state_fields:
+                    attributes.extend(state_fields.keys())
+                    active_user = False
+                else:
+                    active_user = True
+
+                if len(attributes) > 1:
                     conn.search(user_dn, search_filter, ldap3.SUBTREE, attributes=attributes)
 
                     for r in conn.response:
@@ -83,17 +137,26 @@ class LdapAuthenticationBackend(ModelBackend):
                                 active_user = True
                                 break
 
-                        for ef in email_fields:
-                            if ef in attrs and attrs[ef][0]:
-                                email = attrs[ef][0]
+                        for f in email_fields:
+                            if f in attrs and attrs[f][0]:
+                                email = attrs[f][0]
                                 break
 
-                    if active_user:
-                        user = UserModel.objects.create_user(username, email, '')
-                    else:
-                        return None
+                        for f in firstname_fields:
+                            if f in attrs and attrs[f][0]:
+                                firstname = attrs[f][0]
+                                break
+
+                        for f in lastname_fields:
+                            if f in attrs and attrs[f][0]:
+                                lastname = attrs[f][0]
+                                break
+
+                if active_user:
+                    user = UserModel.objects.create_user(
+                        username, email, '', first_name=firstname, last_name=lastname)
                 else:
-                    user = UserModel.objects.create_user(username, '', '')
+                    return None
             else:
                 # Run the default password hasher once to reduce the timing
                 # difference between an existing and a non-existing user (#20760).
@@ -128,6 +191,8 @@ class LdapAuthenticationBackend(ModelBackend):
         return user
 
     def get_user(self, user_id):
+        # TODO maybe we want to lookup into the LDAP server at this level,
+        # like with the auto creation of the user when authenticating.
         UserModel = get_user_model()
         try:
             return UserModel._default_manager.get(pk=user_id)