+++ igdectk/jquery/appsettings.py
@@ -1,0 +1,17 @@
+# -*- coding: utf-8; -*-
+#
+# Copyright (c) 2015 INRA UMR1095 GDEC
+
+"""
+JQuery sub-module/Django-application settings
+"""
+
+# Default settings of the application
+APP_DB_DEFAULT_SETTINGS = {
+}
+
+APP_VERBOSE_NAME = "igdectk.jquery"
+
+APP_SETTINGS_MODEL = 'igdectk.jquery.models.Settings'
+
+APP_VERSION = (1, 0, 0)